requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.1
scipy>=1.6.0
brisque==0.2.0
lpips>=0.1.4
Pillow>=8.0.0
PyYAML>=5.4.0
tqdm>=4.60.0
matplotlib>=3.3.4
